backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
email-validator

openai

google-generativeai

supabase

requests

pymupdf

python-multipart